+++ src/aks-preview/azext_aks_preview/custom.py
@@ -3331,6 +3331,7 @@ def aks_machine_add(
     spot_max_price=float("nan"),
     enable_ultra_ssd=False,
     eviction_policy=None,
+    capacity_reservation_group=None,
 ):
     existedMachine = None
     try: